+++ src/StateSmith/Input/DrawIo/DrawIoDecoder.cs
@@ -1,3 +1,4 @@
+#nullable enable
 
 using System;
 using System.IO;
@@ -5,8 +6,7 @@ using System.IO.Compression;
 using System.Text;
 using System.Xml;
 using System.Text.RegularExpressions;
-
-#nullable enable
+using System.Collections.Generic;
 
 namespace StateSmith.Input.DrawIo;
 
@@ -24,7 +24,7 @@ public class DrawIoDecoder
     /// </summary>
     /// <param name="textReader"></param>
     /// <returns></returns>
-    public static string DecodeSvgToOriginalDiagram(TextReader textReader)
+    public static List<DrawIoDiagramNode> DecodeSvgToOriginalDiagrams(TextReader textReader)
     {
         string mxfileXml = GetMxfileFromSvg(textReader);
 
@@ -39,16 +39,19 @@ public class DrawIoDecoder
     /// </summary>
     /// <param name="mxfileXml"></param>
     /// <returns></returns>
-    public static string GetMxFileDiagramContents(string mxfileXml)
+    public static List<DrawIoDiagramNode> GetMxFileDiagramContents(string mxfileXml)
     {
-        var diagramContents = GetDiagramContentsRaw(mxfileXml);
+        var diagrams = GetDiagramsContentsRaw(mxfileXml);
 
-        if (!IsDiagramContentUncompressed(diagramContents))
+        foreach (var diagram in diagrams)
         {
-            diagramContents = DecompressContent(diagramContents);
+            if (!IsDiagramContentUncompressed(diagram.xml))
+            {
+                diagram.xml = DecompressContent(diagram.xml);
+            }
         }
 
-        return diagramContents;
+        return diagrams;
     }
 
     private static bool IsDiagramContentUncompressed(string diagramContents)
@@ -84,48 +87,60 @@ public class DrawIoDecoder
     /// <param name="mxfileXmlContents"></param>
     /// <returns></returns>
     /// <exception cref="DrawIoException"></exception>
-    public static string GetDiagramContentsRaw(string mxfileXmlContents)
+    public static List<DrawIoDiagramNode> GetDiagramsContentsRaw(string mxfileXmlContents)
     {
+        List<DrawIoDiagramNode> diagrams = new();
         using XmlTextReader reader = new(new StringReader(mxfileXmlContents));
-        
+
         // keep whitespace so that ReadInnerXml contains whitespaces (required for error line numbers)
         reader.WhitespaceHandling = WhitespaceHandling.All;
 
-        ReadAndExpectOpeningXmlTag(reader, "mxfile");
-        ReadAndExpectOpeningXmlTag(reader, "diagram");
-        SkipWhiteSpaceToNextElement(reader);   // skip whitespace to improve error messages
-        var contents = reader.ReadInnerXml();  // advances past diagram closing tag
-        SkipOverAnyAdditionalPages(reader);
-        ExpectCurrentTagToBeClosingXmlTag(reader, "mxfile");
+        ReadOrThrow(reader, "Expecting <mxfile> tag");
+        ExpectOpeningXmlTag(reader, "mxfile");
 
-        return contents.Trim();
+        ReadOrThrow(reader, "Expecting <diagram> tag");
+
+        while (true)
+        {
+            diagrams.Add(ReadDiagramNode(reader));
+            SkipWhiteSpaceToNextElement(reader);
+
+            if (reader.LocalName != "diagram")
+            {
+                break;
+            }
+        }
+
+        ExpectClosingXmlTag(reader, "mxfile");
+
+        return diagrams;
     }
 
     /// <summary>
     /// https://github.com/StateSmith/StateSmith/issues/78
     /// </summary>
-    /// <param name="reader"></param>
-    private static void SkipOverAnyAdditionalPages(XmlTextReader reader)
+    private static DrawIoDiagramNode ReadDiagramNode(XmlTextReader reader)
     {
-        SkipWhiteSpaceToNextElement(reader);
+        DrawIoDiagramNode diagramNode = new();
+        ExpectOpeningXmlTag(reader, "diagram");
+        diagramNode.id = reader.GetAttribute("id") ?? "";
+        diagramNode.name = reader.GetAttribute("name") ?? "";
 
-        while (reader.LocalName == "diagram")
-        {
-            reader.Skip();
-            SkipWhiteSpaceToNextElement(reader);
-        }
+        SkipWhiteSpaceToNextElement(reader);     // skip whitespace to improve error messages
+        diagramNode.xml = reader.ReadInnerXml();  // advances past diagram closing tag
+        diagramNode.xml = diagramNode.xml.Trim();
+        return diagramNode;
     }
 
-    private static void ReadAndExpectOpeningXmlTag(XmlTextReader reader, string tag)
+    private static void ExpectOpeningXmlTag(XmlTextReader reader, string tag)
     {
-        ReadOrThrow(reader, $"Expected opening xml tag `{tag}`");
         if (!reader.IsStartElement(tag))
         {
             throw new DrawIoException($"Invalid draw.io file. Expected opening xml tag `{tag}`, but found `{reader.LocalName}`");
         }
     }
 
-    private static void ExpectCurrentTagToBeClosingXmlTag(XmlTextReader reader, string tag)
+    private static void ExpectClosingXmlTag(XmlTextReader reader, string tag)
     {
         if (reader.NodeType != XmlNodeType.EndElement)
         {
@@ -173,7 +188,7 @@ public class DrawIoDecoder
     {
         using XmlTextReader reader = new(textReader);
 
-        ReadAndExpectOpeningXmlTag(reader, "svg");
+        ExpectOpeningXmlTag(reader, "svg");
 
         var content = reader.GetAttribute("content");